+++ src/components/Flyout/utilities/calculatePosition.ts
@@ -19,25 +19,29 @@ const centerBySize = (originSize: number, targetSize: number) => {
  * Calculate styles for the current position
  */
 const calculatePosition = (
-	args: T.Options & {
+	args: {
 		triggerBounds: DOMRect;
-		flyoutBounds: DOMRect;
-		scopeOffset: Record<"left" | "top", number>;
-	}
+		flyoutBounds: { width: number; height: number };
+		passedContainer?: HTMLElement | null;
+		containerBounds: DOMRect;
+	} & Pick<T.Options, "position" | "rtl" | "width" | "contentGap" | "contentShift">
 ) => {
 	const {
 		triggerBounds,
 		flyoutBounds,
-		scopeOffset,
+		containerBounds,
 		position: passedPosition,
 		rtl,
 		width,
 		contentGap = 0,
 		contentShift = 0,
+		passedContainer,
 	} = args;
 	const isFullWidth = width === "full" || width === "100%";
-	let left = 0;
-	let top = 0;
+	let left: number = 0;
+	let top: number = 0;
+	let bottom: number | null = null;
+	let right: number | null = null;
 
 	let position = passedPosition;
 	if (rtl) position = getRTLPosition(position);
@@ -45,39 +49,52 @@ const calculatePosition = (
 		position = position.includes("top") ? "top" : "bottom";
 	}
 
-	const isHorizontalPosition = position.match(/^(start|end)/);
-	const isVerticalPosition = position.match(/^(top|bottom)/);
+	const isHorizontalPosition = !!position.match(/^(start|end)/);
+	const isVerticalPosition = !!position.match(/^(top|bottom)/);
 
+	// contentGap adds padding to the flyout to make sure it doesn't disapper while moving the mouse to the content
+	// So its width/height is bigger than the visible part of the content
 	const flyoutWidth = flyoutBounds.width + (isHorizontalPosition ? contentGap : 0);
 	const flyoutHeight = flyoutBounds.height + (isVerticalPosition ? contentGap : 0);
 
-	switch (position) {
-		case "bottom":
-		case "top":
-			left = centerBySize(triggerBounds.width, flyoutWidth) + triggerBounds.left + contentShift;
-			break;
+	const triggerHeight = triggerBounds.height;
+	const triggerWidth = triggerBounds.width;
+	const containerY = passedContainer?.scrollTop || 0;
+	const containerX = passedContainer?.scrollLeft || 0;
 
+	const relativeLeft = triggerBounds.left - containerBounds.left + containerX;
+	const relativeTop = triggerBounds.top - containerBounds.top + containerY;
+	const relativeRight = containerBounds.right - triggerBounds.right - containerX;
+	const relativeBottom = containerBounds.bottom - triggerBounds.bottom - containerY;
+
+	switch (position) {
 		case "start":
 		case "start-top":
 		case "start-bottom":
-			left = triggerBounds.left - flyoutWidth;
-
+			right = relativeRight + triggerWidth;
+			left = relativeLeft - flyoutWidth;
 			break;
 
 		case "end":
 		case "end-top":
 		case "end-bottom":
-			left = triggerBounds.right;
+			left = relativeLeft + triggerWidth;
+			break;
+
+		case "bottom":
+		case "top":
+			left = relativeLeft + centerBySize(triggerWidth, flyoutWidth) + contentShift;
 			break;
 
 		case "top-start":
 		case "bottom-start":
-			left = triggerBounds.left + contentShift + contentShift;
+			left = relativeLeft + contentShift;
 			break;
 
 		case "top-end":
 		case "bottom-end":
-			left = triggerBounds.right - flyoutWidth + contentShift;
+			right = relativeRight - contentShift;
+			left = relativeLeft + triggerWidth - flyoutWidth + contentShift;
 			break;
 
 		default:
@@ -88,43 +105,37 @@ const calculatePosition = (
 		case "top":
 		case "top-start":
 		case "top-end":
-			top = triggerBounds.top - flyoutHeight;
+			bottom = relativeBottom + triggerHeight;
+			top = relativeTop - flyoutHeight;
 			break;
 
 		case "bottom":
 		case "bottom-start":
 		case "bottom-end":
-			top = triggerBounds.bottom;
+			top = relativeTop + triggerHeight;
 			break;
 
 		case "start":
 		case "end":
-			top = centerBySize(triggerBounds.height, flyoutHeight) + triggerBounds.top + contentShift;
+			top = relativeTop + centerBySize(triggerHeight, flyoutHeight) + contentShift;
 			break;
 
 		case "start-top":
 		case "end-top":
-			top = triggerBounds.top + contentShift;
+			top = relativeTop + contentShift;
 			break;
 
 		case "start-bottom":
 		case "end-bottom":
-			top = triggerBounds.bottom - flyoutHeight + contentShift;
+			bottom = relativeBottom - contentShift;
+			top = relativeTop + triggerHeight - flyoutHeight + contentShift;
 			break;
 
 		default:
 			break;
 	}
 
-	if (top === undefined || left === undefined) {
-		throw Error(`[Reshaped, flyout]: ${position} position is not valid`);
-	}
-
-	top = Math.round(top + (window.scrollY || 0) - scopeOffset.top);
-	left = Math.round(left + (window.scrollX || 0) - scopeOffset.left);
-	let widthStyle = Math.ceil(flyoutWidth);
-	const height = Math.ceil(flyoutHeight);
-
+	let widthStyle;
 	if (isFullWidth) {
 		left = SCREEN_OFFSET;
 		widthStyle = window.innerWidth - SCREEN_OFFSET * 2;
@@ -132,9 +143,26 @@ const calculatePosition = (
 		widthStyle = triggerBounds.width;
 	}
 
-	const styles = { left, top, width: widthStyle, height };
+	const translateX = right !== null ? -right : left;
+	const translateY = bottom !== null ? -bottom : top;
 
-	return { styles, position, scopeOffset };
+	return {
+		position,
+		styles: {
+			width: widthStyle,
+			left: right === null ? 0 : undefined,
+			right: right === null ? undefined : 0,
+			top: bottom === null ? 0 : undefined,
+			bottom: bottom === null ? undefined : 0,
+			transform: `translate(${translateX}px, ${translateY}px)`,
+		},
+		boundaries: {
+			left,
+			top,
+			height: Math.ceil(flyoutHeight),
+			width: widthStyle ?? Math.ceil(flyoutWidth),
+		},
+	};
 };
 
 export default calculatePosition;