+++ src/utilities/dom/index.ts
@@ -1,4 +1,4 @@
 export { getRectFromCoordinates } from "./flyout";
 export { getShadowRoot } from "./shadowDom";
 export { enableUserSelect, disableUserSelect } from "./userSelect";
-export { findParent, findClosestRenderContainer } from "./find";
+export { findParent, findClosestScrollableContainer, findClosestPositionContainer } from "./find";